PS-354 Keep both old & new implementation of table upload, switch using feature flag
