+++ tests/Writer/StorageApiTableWriterV1Test.php
@@ -7,18 +7,18 @@ use Keboola\OutputMapping\Exception\OutputOperationException;
 use Keboola\OutputMapping\Staging\StrategyFactory;
 use Keboola\OutputMapping\Tests\Writer\BaseWriterTest;
 use Keboola\OutputMapping\Tests\Writer\CreateBranchTrait;
-use Keboola\OutputMapping\Writer\FileWriter;
-use Keboola\OutputMapping\Writer\TableWriter;
+use Keboola\OutputMapping\Writer\Table\TableWriterV1;
 use Keboola\StorageApi\Client;
 use Keboola\StorageApi\Metadata;
-use Keboola\StorageApi\Options\FileUploadOptions;
-use Keboola\StorageApi\Options\ListFilesOptions;
 use Keboola\StorageApi\TableExporter;
 use Keboola\StorageApiBranch\ClientWrapper;
 use Psr\Log\NullLogger;
 use Psr\Log\Test\TestLogger;
 
-class StorageApiWriterTest extends BaseWriterTest
+/**
+ * @group tableWriterV1
+ */
+class StorageApiTableWriterV1Test extends BaseWriterTest
 {
     use CreateBranchTrait;
 
@@ -39,324 +39,6 @@ class StorageApiWriterTest extends BaseWriterTest
         $this->clientWrapper->getBasicClient()->createBucket('output-mapping-default-test', 'out');
     }
 
-    public function testWriteBasicFiles()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-        file_put_contents($root . "/upload/file2", "test");
-        file_put_contents(
-            $root . "/upload/file2.manifest",
-            "{\"tags\": [\"output-mapping-test\", \"xxx\"],\"is_public\": false}"
-        );
-        file_put_contents($root . "/upload/file3", "test");
-        file_put_contents($root . "/upload/file3.manifest", "{\"tags\": [\"output-mapping-test\"],\"is_public\": true}");
-
-        $systemMetadata = [
-            "componentId" => "testComponent",
-            "configurationId" => "metadata-write-test",
-            "configurationRowId" => "12345",
-            "branchId" => "1234",
-            "runId" => "999",
-        ];
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test"]
-            ],
-            [
-                "source" => "file2",
-                "tags" => ["output-mapping-test", "another-tag"],
-                "is_public" => true
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-
-        $writer->uploadFiles('/upload', ["mapping" => $configs], $systemMetadata, StrategyFactory::LOCAL);
-        sleep(1);
-
-        $options = new ListFilesOptions();
-        $options->setTags(["output-mapping-test"]);
-        $files = $this->clientWrapper->getBasicClient()->listFiles($options);
-        $this->assertCount(3, $files);
-
-        $file1 = $file2 = $file3 = null;
-        foreach ($files as $file) {
-            if ($file["name"] == 'file1') {
-                $file1 = $file;
-            }
-            if ($file["name"] == 'file2') {
-                $file2 = $file;
-            }
-            if ($file["name"] == 'file3') {
-                $file3 = $file;
-            }
-        }
-
-        $expectedTags = [
-            'output-mapping-test',
-            'componentId: testComponent',
-            'configurationId: metadata-write-test',
-            'configurationRowId: 12345',
-            'branchId: 1234',
-            'runId: 999',
-        ];
-        $expectedFile2Tags = array_merge($expectedTags, ['another-tag']);
-
-        $this->assertNotNull($file1);
-        $this->assertNotNull($file2);
-        $this->assertNotNull($file3);
-        $this->assertEquals(4, $file1["sizeBytes"]);
-        $this->assertEquals($expectedTags, $file1["tags"]);
-        $this->assertEquals(sort($expectedFile2Tags), sort($file2["tags"]));
-        $this->assertEquals($expectedTags, $file3["tags"]);
-        $this->assertFalse($file1["isPublic"]);
-        $this->assertTrue($file2["isPublic"]);
-        $this->assertTrue($file3["isPublic"]);
-    }
-
-    public function testWriteFilesOutputMapping()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test"]
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-
-        $writer->uploadFiles('/upload', ["mapping" => $configs], [], StrategyFactory::LOCAL);
-        sleep(1);
-
-        $options = new ListFilesOptions();
-        $options->setTags(["output-mapping-test"]);
-        $files = $this->clientWrapper->getBasicClient()->listFiles($options);
-        $this->assertCount(1, $files);
-
-        $file1 = null;
-        foreach ($files as $file) {
-            if ($file["name"] == 'file1') {
-                $file1 = $file;
-            }
-        }
-
-        $this->assertNotNull($file1);
-        $this->assertEquals(4, $file1['sizeBytes']);
-        $this->assertEquals(['output-mapping-test'], $file1['tags']);
-    }
-
-    public function testWriteFilesOutputMappingDevMode()
-    {
-        $this->clearFileUploads(['dev-123-output-mapping-test']);
-        $this->clientWrapper = new ClientWrapper(
-            new Client([
-                'url' => STORAGE_API_URL,
-                'token' => STORAGE_API_TOKEN_MASTER,
-                'backoffMaxTries' => 1,
-                'jobPollRetryDelay' => function () {
-                    return 1;
-                },
-            ]),
-            null,
-            null
-        );
-        $branchId = $this->createBranch($this->clientWrapper, 'dev-123');
-        $this->clientWrapper->setBranchId($branchId);
-
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test"]
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-
-        $systemMetadata = [
-            'componentId' => 'testComponent',
-            'configurationId' => 'metadata-write-test',
-            'configurationRowId' => '12345',
-            'branchId' => $branchId,
-            'runId' => '999',
-        ];
-
-        $writer->uploadFiles('/upload', ['mapping' => $configs], $systemMetadata, StrategyFactory::LOCAL);
-        sleep(1);
-
-        $options = new ListFilesOptions();
-        $options->setTags([sprintf('%s-output-mapping-test', $branchId)]);
-        $files = $this->clientWrapper->getBasicClient()->listFiles($options);
-        $this->assertCount(1, $files);
-
-        $file1 = null;
-        foreach ($files as $file) {
-            if ($file["name"] == 'file1') {
-                $file1 = $file;
-            }
-        }
-
-        $expectedTags = [
-            sprintf('%s-output-mapping-test', $branchId),
-            sprintf('%s-componentId: testComponent', $branchId),
-            sprintf('%s-configurationId: metadata-write-test', $branchId),
-            sprintf('%s-configurationRowId: 12345', $branchId),
-            sprintf('%s-branchId: %s', $branchId, $branchId),
-            sprintf('%s-runId: 999', $branchId),
-        ];
-
-        $this->assertNotNull($file1);
-        $this->assertEquals(4, $file1["sizeBytes"]);
-        $this->assertEquals($expectedTags, $file1['tags']);
-    }
-
-    public function testWriteFilesOutputMappingAndManifest()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-        file_put_contents(
-            $root . "/upload/file1.manifest",
-            "{\"tags\": [\"output-mapping-test\", \"xxx\"],\"is_public\": true}"
-        );
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test", "yyy"],
-                "is_public" => false
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-        $writer->uploadFiles('upload', ["mapping" => $configs], ['componentId' => 'foo'], StrategyFactory::LOCAL);
-        sleep(1);
-
-        $options = new ListFilesOptions();
-        $options->setTags(["output-mapping-test"]);
-        $files = $this->clientWrapper->getBasicClient()->listFiles($options);
-        $this->assertCount(1, $files);
-
-        $file1 = null;
-        foreach ($files as $file) {
-            if ($file["name"] == 'file1') {
-                $file1 = $file;
-            }
-        }
-
-        $this->assertNotNull($file1);
-        $this->assertEquals(4, $file1["sizeBytes"]);
-        $this->assertEquals(["output-mapping-test", "yyy", "componentId: foo"], $file1["tags"]);
-        $this->assertFalse($file1["isPublic"]);
-    }
-
-    public function testWriteFilesInvalidJson()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-        file_put_contents($root . "/upload/file1.manifest", "this is not at all a {valid} json");
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test", "yyy"],
-                "is_public" => false
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-        $writer->setFormat('json');
-        try {
-            $writer->uploadFiles('/upload', ['mapping' => $configs], [], StrategyFactory::LOCAL);
-            $this->fail('Invalid manifest must raise exception.');
-        } catch (InvalidOutputException $e) {
-            $this->assertContains('json', $e->getMessage());
-        }
-    }
-
-    public function testWriteFilesInvalidYaml()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-        file_put_contents($root . "/upload/file1.manifest", "\tthis is not \n\t \tat all a {valid} json");
-
-        $configs = [
-            [
-                "source" => "file1",
-                "tags" => ["output-mapping-test", "yyy"],
-                "is_public" => false
-            ]
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-        $writer->setFormat('json');
-        try {
-            $writer->uploadFiles('upload', ['mapping' => $configs], [], StrategyFactory::LOCAL);
-            $this->fail('Invalid manifest must raise exception.');
-        } catch (InvalidOutputException $e) {
-            $this->assertContains('json', $e->getMessage());
-        }
-    }
-
-    public function testWriteFilesOutputMappingMissing()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/file1", "test");
-        file_put_contents(
-            $root . "/upload/file1.manifest",
-            "{\"tags\": [\"output-mapping-test-xxx\"],\"is_public\": true}"
-        );
-
-        $configs = [
-            [
-                "source" => "file2",
-                "tags" => ["output-mapping-test"],
-                "is_public" => false
-            ]
-        ];
-        $writer = new FileWriter($this->getStagingFactory());
-        try {
-            $writer->uploadFiles('upload', ['mapping' => $configs], [], StrategyFactory::LOCAL);
-            $this->fail('Missing file must fail');
-        } catch (InvalidOutputException $e) {
-            $this->assertContains("File 'file2' not found", $e->getMessage());
-            $this->assertEquals(404, $e->getCode());
-        }
-    }
-
-    public function testWriteFilesOrphanedManifest()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents(
-            $root . "/upload/file1.manifest",
-            "{\"tags\": [\"output-mapping-test-xxx\"],\"is_public\": true}"
-        );
-        $writer = new FileWriter($this->getStagingFactory());
-        try {
-            $writer->uploadFiles('/upload', [], [], StrategyFactory::LOCAL);
-            $this->fail('Orphaned manifest must cause exception.');
-        } catch (InvalidOutputException $e) {
-            $this->assertContains("Found orphaned file manifest: 'file1.manifest'", $e->getMessage());
-        }
-    }
-
-    public function testWriteFilesNoComponentId()
-    {
-        $writer = new FileWriter($this->getStagingFactory());
-        try {
-            $writer->uploadFiles('/upload', [], ['configurationId' => '123'], StrategyFactory::LOCAL);
-            $this->fail('Missing componentId must cause exception.');
-        } catch (OutputOperationException $e) {
-            $this->assertContains('Component Id must be set', $e->getMessage());
-        }
-    }
-
     public function testWriteTableOutputMapping()
     {
         $root = $this->tmp->getTmpFolder();
@@ -374,7 +56,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -423,7 +105,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $tokenInfo['owner']['features'][] = 'tag-staging-files';
         $client->method('verifyToken')->willReturn($tokenInfo);
         $clientWrapper = new ClientWrapper($client, null, new NullLogger(), '');
-        $writer = new TableWriter($this->getStagingFactory($clientWrapper));
+        $writer = new TableWriterV1($this->getStagingFactory($clientWrapper));
 
         $tableQueue =  $writer->uploadTables('upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -479,7 +161,7 @@ class StorageApiWriterTest extends BaseWriterTest
                 "destination" => "out.c-output-mapping-test.table21a"
             ]
         ];
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             '/upload',
             ['mapping' => $configs],
@@ -516,7 +198,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -544,7 +226,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -567,7 +249,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         try {
             $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
@@ -596,7 +278,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -605,7 +287,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
         $this->assertCount(1, $tables);
         $this->assertEquals('out.c-output-mapping-test.table', $tables[0]["id"]);
-        $this->assertEquals(['Id'], $tables[0]["primaryKey"]);
+        $this->assertEquals([], $tables[0]["primaryKey"]);
     }
 
     public function testWriteTableManifest()
@@ -620,7 +302,7 @@ class StorageApiWriterTest extends BaseWriterTest
             "{\"destination\": \"out.c-output-mapping-test.table3a\",\"primary_key\": [\"Id\",\"Name\"]}"
         );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -644,7 +326,7 @@ class StorageApiWriterTest extends BaseWriterTest
             "{\"destination\": \"out.c-output-mapping-test.table3\",\"primary_key\": \"Id\"}"
         );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
             $this->fail('Invalid table manifest must cause exception');
@@ -665,7 +347,7 @@ class StorageApiWriterTest extends BaseWriterTest
             "{\"destination\": \"out.c-output-mapping-default-test.table3c\",\"delimiter\": \"\\t\",\"enclosure\": \"'\"}"
         );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -698,7 +380,7 @@ class StorageApiWriterTest extends BaseWriterTest
             "{\"destination\": \"out.c-output-mapping-redshift-test.table3d\",\"delimiter\": \"\\t\",\"enclosure\": \"'\"}"
         );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
@@ -725,12 +407,12 @@ class StorageApiWriterTest extends BaseWriterTest
             $root . DIRECTORY_SEPARATOR . "upload/table.csv.manifest",
             "{\"destination\": \"out.c-output-mapping-test.table3e\",\"primary_key\": [\"Id\", \"Name\"]}"
         );
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
             $this->fail("Orphaned manifest must fail");
         } catch (InvalidOutputException $e) {
-            $this->assertContains('Found orphaned table manifest: "table.csv.manifest"', $e->getMessage());
+            $this->assertContains("Found orphaned table manifest: 'table.csv.manifest'", $e->getMessage());
         }
     }
 
@@ -744,19 +426,21 @@ class StorageApiWriterTest extends BaseWriterTest
                 "destination" => "out.c-output-mapping-test.table81"
             ]
         ];
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
             $this->fail("Missing table file must fail");
         } catch (InvalidOutputException $e) {
-            $this->assertContains('Table source "table81.csv" not found', $e->getMessage());
+            $this->assertContains("Table source 'table81.csv' not found", $e->getMessage());
             $this->assertEquals(404, $e->getCode());
         }
     }
 
     public function testWriteTableMetadataMissing()
     {
-        $writer = new TableWriter($this->getStagingFactory());
+        $root = $this->tmp->getTmpFolder();
+
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables('/upload', [], [], 'local');
             self::fail("Missing metadata must fail.");
@@ -765,17 +449,42 @@ class StorageApiWriterTest extends BaseWriterTest
         }
     }
 
-    public function testWriteTableBareFails()
+    public function testWriteTableBare()
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/out.c-output-mapping-test.table4.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
-        $this->expectException(InvalidOutputException::class);
-        $this->expectExceptionMessage('Failed to resolve destination for output table "out.c-output-mapping-test.table4.csv".');
+        $tableQueue =  $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
+        $jobIds = $tableQueue->waitForAll();
+        $this->assertCount(1, $jobIds);
 
-        $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
+        $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
+        $this->assertCount(1, $tables);
+
+        $this->assertEquals('out.c-output-mapping-test.table4', $tables[0]["id"]);
+        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('out.c-output-mapping-test.table4');
+        $this->assertEquals(["Id", "Name"], $tableInfo["columns"]);
+    }
+
+    public function testWriteTableBareWithoutSuffix()
+    {
+        $root = $this->tmp->getTmpFolder();
+        file_put_contents($root . "/upload/out.c-output-mapping-test.table4", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
+
+        $writer = new TableWriterV1($this->getStagingFactory());
+
+        $tableQueue =  $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
+        $jobIds = $tableQueue->waitForAll();
+        $this->assertCount(1, $jobIds);
+
+        $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
+        $this->assertCount(1, $tables);
+
+        $this->assertEquals('out.c-output-mapping-test.table4', $tables[0]["id"]);
+        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('out.c-output-mapping-test.table4');
+        $this->assertEquals(["Id", "Name"], $tableInfo["columns"]);
     }
 
     public function testWriteTableIncrementalWithDeleteDefault()
@@ -794,7 +503,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -841,7 +550,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables('/upload', ["mapping" => $configs], ['componentId' => 'foo'], 'local');
         $jobIds = $tableQueue->waitForAll();
@@ -870,93 +579,67 @@ class StorageApiWriterTest extends BaseWriterTest
         $this->assertEquals('test', $table[2]['Name']);
     }
 
-    public function testTagProcessedFiles()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/test", "test");
-
-        $id1 = $this->clientWrapper->getBasicClient()->uploadFile(
-            $root . "/upload/test",
-            (new FileUploadOptions())->setTags(["output-mapping-test"])
-        );
-        $id2 = $this->clientWrapper->getBasicClient()->uploadFile(
-            $root . "/upload/test",
-            (new FileUploadOptions())->setTags(["output-mapping-test"])
-        );
-        sleep(1);
-
-        $writer = new FileWriter($this->getStagingFactory());
-        $configuration = [["tags" => ["output-mapping-test"], "processed_tags" => ['downloaded']]];
-        $writer->tagFiles($configuration);
-
-        $file = $this->clientWrapper->getBasicClient()->getFile($id1);
-        $this->assertTrue(in_array('downloaded', $file['tags']));
-        $file = $this->clientWrapper->getBasicClient()->getFile($id2);
-        $this->assertTrue(in_array('downloaded', $file['tags']));
-    }
-
-    public function testTagBranchProcessedFiles()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/test", "test");
-
-        $id1 = $this->clientWrapper->getBasicClient()->uploadFile(
-            $root . "/upload/test",
-            (new FileUploadOptions())->setTags(["output-mapping-test"])
-        );
-        $id2 = $this->clientWrapper->getBasicClient()->uploadFile(
-            $root . "/upload/test",
-            (new FileUploadOptions())->setTags(["12345-output-mapping-test"])
-        );
-        sleep(1);
-        // set it to use a branch
-        $this->initClient('12345');
-
-        $writer = new FileWriter($this->getStagingFactory());
-        $configuration = [["tags" => ["output-mapping-test"], "processed_tags" => ['downloaded']]];
-        $writer->tagFiles($configuration);
-
-        // first file shouldn't be marked as processed because a branch file exists
-        $file1 = $this->clientWrapper->getBasicClient()->getFile($id1);
-        $this->assertTrue(!in_array('12345-downloaded', $file1['tags']));
-        $file2 = $this->clientWrapper->getBasicClient()->getFile($id2);
-        $this->assertTrue(in_array('12345-downloaded', $file2['tags']));
-        $this->assertTrue(in_array('12345-output-mapping-test', $file2['tags']));
-    }
-
     public function testWriteTableToDefaultBucket()
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table71.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
-        file_put_contents($root . "/upload/table71.csv.manifest", '{}');
+        file_put_contents($root . "/upload/table72.csv", "\"Id\",\"Name2\"\n\"test2\",\"test2\"\n");
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
+
+        $tableQueue =  $writer->uploadTables('/upload', ["bucket" => "in.c-output-mapping-test"], ['componentId' => 'foo'], 'local');
+        $jobIds = $tableQueue->waitForAll();
+        $this->assertCount(2, $jobIds);
+        $this->assertEquals(2, $tableQueue->getTaskCount());
+
+        $tables = $this->clientWrapper->getBasicClient()->listTables("in.c-output-mapping-test");
+        $this->assertCount(2, $tables);
+
+        $this->assertEquals('in.c-output-mapping-test.table71', $tables[0]["id"]);
+        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('in.c-output-mapping-test.table71');
+        $this->assertEquals(["Id", "Name"], $tableInfo["columns"]);
+
+        $this->assertEquals('in.c-output-mapping-test.table72', $tables[1]["id"]);
+        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('in.c-output-mapping-test.table72');
+        $this->assertEquals(["Id", "Name2"], $tableInfo["columns"]);
+    }
+
+    public function testWriteTableBareWithDefaultBucket()
+    {
+        $root = $this->tmp->getTmpFolder();
+        file_put_contents($root . "/upload/table5.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
+
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables(
             '/upload',
-            ['bucket' => 'in.c-output-mapping-test'],
+            ['bucket' => 'out.c-output-mapping-test'],
             ['componentId' => 'foo'],
             'local'
         );
         $jobIds = $tableQueue->waitForAll();
-        $this->assertCount(1, $jobIds);
-        $this->assertEquals(1, $tableQueue->getTaskCount());
-
-        $tables = $this->clientWrapper->getBasicClient()->listTables("in.c-output-mapping-test");
+        $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
         $this->assertCount(1, $tables);
+        $this->assertCount(1, $jobIds);
 
-        $this->assertEquals('in.c-output-mapping-test.table71', $tables[0]["id"]);
-        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('in.c-output-mapping-test.table71');
+        $this->assertEquals('out.c-output-mapping-test.table5', $tables[0]["id"]);
+        $tableInfo = $this->clientWrapper->getBasicClient()->getTable('out.c-output-mapping-test.table5');
         $this->assertEquals(["Id", "Name"], $tableInfo["columns"]);
     }
 
     public function testWriteTableManifestWithDefaultBucket()
     {
         $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . "/upload/table6.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
-        file_put_contents($root . "/upload/table6.csv.manifest", '{"primary_key": ["Id", "Name"]}');
+        file_put_contents(
+            $root . "/upload/table6.csv",
+            "\"Id\",\"Name\"\n\"test\",\"test\"\n"
+        );
+        file_put_contents(
+            $root . "/upload/table6.csv.manifest",
+            "{\"primary_key\": [\"Id\", \"Name\"]}"
+        );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
 
         $tableQueue =  $writer->uploadTables(
             '/upload',
@@ -973,18 +656,47 @@ class StorageApiWriterTest extends BaseWriterTest
         $this->assertEquals(["Id", "Name"], $tables[0]["primaryKey"]);
     }
 
+    public function testWriteTableOutputMappingWithDefaultBucket()
+    {
+        $root = $this->tmp->getTmpFolder();
+        file_put_contents($root . "/upload/table7.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n\"aabb\",\"ccdd\"\n");
+
+        $configs = [
+            [
+                "source" => "table7.csv",
+                "destination" => "table7"
+            ]
+        ];
+
+        $writer = new TableWriterV1($this->getStagingFactory());
+
+        $tableQueue =  $writer->uploadTables(
+            '/upload',
+            ["mapping" => $configs, 'bucket' => 'out.c-output-mapping-test'],
+            ['componentId' => 'foo'],
+            'local'
+        );
+        $jobIds = $tableQueue->waitForAll();
+        $this->assertCount(1, $jobIds);
+
+        $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
+        $this->assertCount(1, $tables);
+        $this->assertEquals('out.c-output-mapping-test.table7', $tables[0]["id"]);
+    }
+
     public function testWriteManifestWithoutDestination()
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table8.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
         file_put_contents($root . "/upload/table8.csv.manifest", "{\"primary_key\": [\"Id\", \"Name\"]}");
 
-        $writer = new TableWriter($this->getStagingFactory());
-
-        $this->expectException(InvalidOutputException::class);
-        $this->expectExceptionMessage('Failed to resolve destination for output table "table8.csv".');
-
-        $writer->uploadTables('upload', ['mapping' => []], ['componentId' => 'foo'], 'local');
+        $writer = new TableWriterV1($this->getStagingFactory());
+        try {
+            $writer->uploadTables('upload', ["mapping" => []], ['componentId' => 'foo'], 'local');
+            $this->fail("Empty destination with invalid table name must cause exception.");
+        } catch (InvalidOutputException $e) {
+            $this->assertContains('valid table identifier', $e->getMessage());
+        }
     }
 
     public function testWriteTableOutputMappingWithPk()
@@ -992,7 +704,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table16.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             'upload',
             [
@@ -1018,7 +730,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table15.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $writer->uploadTables(
             'upload',
             [
@@ -1034,7 +746,7 @@ class StorageApiWriterTest extends BaseWriterTest
             'local'
         );
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             '/upload',
             [
@@ -1060,7 +772,7 @@ class StorageApiWriterTest extends BaseWriterTest
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table14.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             'upload',
             [
@@ -1080,7 +792,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $tableInfo = $this->clientWrapper->getBasicClient()->getTable("out.c-output-mapping-test.table14");
         $this->assertEquals(["Id"], $tableInfo["primaryKey"]);
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables(
                 'upload',
@@ -1109,7 +821,7 @@ class StorageApiWriterTest extends BaseWriterTest
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/table13.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             'upload',
             [
@@ -1129,7 +841,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $tableInfo = $this->clientWrapper->getBasicClient()->getTable("out.c-output-mapping-test.table13");
         $this->assertEquals(["Id"], $tableInfo["primaryKey"]);
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         try {
             $writer->uploadTables(
                 'upload',
@@ -1160,7 +872,7 @@ class StorageApiWriterTest extends BaseWriterTest
         file_put_contents($root . "/upload/table12.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
 
         $testLogger = new TestLogger();
-        $writer = new TableWriter($this->getStagingFactory(null, 'json', $testLogger));
+        $writer = new TableWriterV1($this->getStagingFactory(null, 'json', $testLogger));
         $writer->uploadTables(
             'upload',
             [
@@ -1178,7 +890,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $tableInfo = $this->clientWrapper->getBasicClient()->getTable("out.c-output-mapping-test.table12");
         $this->assertEquals([], $tableInfo["primaryKey"]);
 
-        $writer = new TableWriter($this->getStagingFactory(null, 'json', $testLogger));
+        $writer = new TableWriterV1($this->getStagingFactory(null, 'json', $testLogger));
         $tableQueue =  $writer->uploadTables(
             'upload',
             [
@@ -1207,7 +919,7 @@ class StorageApiWriterTest extends BaseWriterTest
         file_put_contents($root . "/upload/table11.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
 
         $testLogger = new TestLogger();
-        $writer = new TableWriter($this->getStagingFactory(null, 'json', $testLogger));
+        $writer = new TableWriterV1($this->getStagingFactory(null, 'json', $testLogger));
         $tableQueue =  $writer->uploadTables(
             'upload',
             [
@@ -1225,7 +937,7 @@ class StorageApiWriterTest extends BaseWriterTest
         $jobIds = $tableQueue->waitForAll();
         $this->assertCount(1, $jobIds);
 
-        $writer = new TableWriter($this->getStagingFactory(null, 'json', $testLogger));
+        $writer = new TableWriterV1($this->getStagingFactory(null, 'json', $testLogger));
         file_put_contents(
             $root . "/upload/table11.csv.manifest",
             '{"destination": "out.c-output-mapping-test.table11","primary_key": [""]}'
@@ -1246,16 +958,8 @@ class StorageApiWriterTest extends BaseWriterTest
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/out.c-output-mapping-test.table10.csv", "\"Id\",\"Name\"\n\"test\",\"test\"\n");
-        $writer = new TableWriter($this->getStagingFactory());
-        $configuration = [
-            'mapping' => [
-                [
-                    'source' => 'out.c-output-mapping-test.table10.csv',
-                    'destination' => 'out.c-output-mapping-test.table10',
-                ]
-            ]
-        ];
-        $tableQueue = $writer->uploadTables('upload', $configuration, ['componentId' => 'foo'], 'local');
+        $writer = new TableWriterV1($this->getStagingFactory());
+        $tableQueue = $writer->uploadTables('upload', [], ['componentId' => 'foo'], 'local');
         $tableQueue->waitForAll();
 
         $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
@@ -1266,18 +970,13 @@ class StorageApiWriterTest extends BaseWriterTest
         $this->assertEquals(["Id", "Name"], $tableInfo["columns"]);
 
         file_put_contents($root . "/upload/out.c-output-mapping-test.table10.csv", "\"foo\",\"bar\"\n\"baz\",\"bat\"\n");
-        $writer = new TableWriter($this->getStagingFactory());
-        $configuration = [
-            'mapping' => [
-                [
-                    'source' => 'out.c-output-mapping-test.table10.csv',
-                    'destination' => 'out.c-output-mapping-test.table10',
-                    'columns' => ['Boing', 'Tschak']
-                ]
-            ]
-        ];
-        $tableQueue = $writer->uploadTables('upload', $configuration, ['componentId' => 'foo'], 'local');
-
+        $writer = new TableWriterV1($this->getStagingFactory());
+        $tableQueue =  $writer->uploadTables(
+            'upload',
+            ["mapping" => [["source" => "out.c-output-mapping-test.table10.csv", "columns" => ["Boing", "Tschak"]]]],
+            ['componentId' => 'foo'],
+            'local'
+        );
         $this->expectException(InvalidOutputException::class);
         $this->expectExceptionMessageRegExp('/Some columns are missing in the csv file. Missing columns: id,name./i');
         $tableQueue->waitForAll();
@@ -1288,21 +987,13 @@ class StorageApiWriterTest extends BaseWriterTest
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/out.c-output-mapping-test.table10a.csv", "\"id\",\"name\"\n\"test\",\"test\"\n");
         file_put_contents($root . "/upload/out.c-output-mapping-test.table10b.csv", "\"foo\",\"bar\"\n\"baz\",\"bat\"\n");
-        $writer = new TableWriter($this->getStagingFactory());
-        $configuration = [
-            'mapping' => [
-                [
-                    'source' => 'out.c-output-mapping-test.table10a.csv',
-                    'destination' => 'out.c-output-mapping-test.table10a',
-                ],
-                [
-                    'source' => 'out.c-output-mapping-test.table10b.csv',
-                    'destination' => 'out.c-output-mapping-test.table10b',
-                ],
-            ]
-        ];
-
-        $tableQueue =  $writer->uploadTables('upload', $configuration, ['componentId' => 'foo'], 'local');
+        $writer = new TableWriterV1($this->getStagingFactory());
+        $tableQueue =  $writer->uploadTables(
+            'upload',
+            ["mapping" => [["source" => "out.c-output-mapping-test.table10a.csv"], ["source" => "out.c-output-mapping-test.table10b.csv"]]],
+            ['componentId' => 'foo'],
+            'local'
+        );
         $tableQueue->waitForAll();
 
         $tables = $this->clientWrapper->getBasicClient()->listTables("out.c-output-mapping-test");
@@ -1312,22 +1003,18 @@ class StorageApiWriterTest extends BaseWriterTest
         $tableInfo = $this->clientWrapper->getBasicClient()->getTable('out.c-output-mapping-test.table10b');
         $this->assertEquals(["foo", "bar"], $tableInfo["columns"]);
 
-        $writer = new TableWriter($this->getStagingFactory());
-        $configuration = [
-            'mapping' => [
+        $writer = new TableWriterV1($this->getStagingFactory());
+        $tableQueue =  $writer->uploadTables(
+            'upload',
+            ["mapping" =>
                 [
-                    'source' => 'out.c-output-mapping-test.table10a.csv',
-                    'destination' => 'out.c-output-mapping-test.table10a',
-                    'columns' => ['Boing', 'Tschak'],
-                ],
-                [
-                    'source' => 'out.c-output-mapping-test.table10b.csv',
-                    'destination' => 'out.c-output-mapping-test.table10b',
-                    'columns' => ['bum', 'tschak'],
-                ],
+                    ["source" => "out.c-output-mapping-test.table10a.csv", "columns" => ["Boing", "Tschak"]],
+                    ["source" => "out.c-output-mapping-test.table10b.csv", "columns" => ["bum", "tschak"]]
+                ]
             ],
-        ];
-        $tableQueue =  $writer->uploadTables('upload', $configuration, ['componentId' => 'foo'], 'local');
+            ['componentId' => 'foo'],
+            'local'
+        );
         try {
             $tableQueue->waitForAll();
             $this->fail("Must raise exception");
@@ -1351,18 +1038,10 @@ class StorageApiWriterTest extends BaseWriterTest
     {
         $root = $this->tmp->getTmpFolder();
         file_put_contents($root . "/upload/out.c-output-mapping-test.table10.csv", "\"Id\",\"Name\"\r\"test\",\"test\"\r");
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $this->expectException(InvalidOutputException::class);
         $this->expectExceptionMessage('Invalid line break. Please use unix \n or win \r\n line breaks.');
-
-        $writer->uploadTables('/upload', [
-            'mapping' => [
-                [
-                    'source' => 'out.c-output-mapping-test.table10.csv',
-                    'destination' => 'out.c-output-mapping-test.table10'
-                ]
-            ]
-        ], ['componentId' => 'foo'], 'local');
+        $writer->uploadTables('/upload', [], ['componentId' => 'foo'], 'local');
     }
 
     public function testWriteTableExistingBucketDevModeNoDev()
@@ -1392,7 +1071,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             '/upload',
             ['mapping' => $configs],
@@ -1410,8 +1089,8 @@ class StorageApiWriterTest extends BaseWriterTest
                 $metadata->deleteBucketMetadata($bucketId, $metadatum['id']);
             }
         }
-        self::expectException(InvalidOutputException::class);
-        self::expectExceptionMessage(sprintf(
+        $this->expectException(InvalidOutputException::class);
+        $this->expectExceptionMessage(sprintf(
             'Trying to create a table in the development bucket ' .
             '"out.c-%s-output-mapping-test" on branch "dev-123" (ID "%s"), but the bucket is not assigned ' .
             'to any development branch.',
@@ -1453,7 +1132,7 @@ class StorageApiWriterTest extends BaseWriterTest
             ]
         ];
 
-        $writer = new TableWriter($this->getStagingFactory());
+        $writer = new TableWriterV1($this->getStagingFactory());
         $tableQueue =  $writer->uploadTables(
             '/upload',
             ['mapping' => $configs],
@@ -1482,8 +1161,8 @@ class StorageApiWriterTest extends BaseWriterTest
             }
         }
 
-        self::expectException(InvalidOutputException::class);
-        self::expectExceptionMessage(sprintf(
+        $this->expectException(InvalidOutputException::class);
+        $this->expectExceptionMessage(sprintf(
             'Trying to create a table in the development bucket ' .
             '"out.c-%s-output-mapping-test" on branch "dev-123" (ID "%s"). ' .
             'The bucket metadata marks it as assigned to branch with ID "12345".',
@@ -1496,58 +1175,5 @@ class StorageApiWriterTest extends BaseWriterTest
             ['componentId' => 'foo'],
             StrategyFactory::LOCAL
         );
-    }
-
-    public function testTableFiles()
-    {
-        $root = $this->tmp->getTmpFolder();
-        file_put_contents($root . '/upload/file', 'test');
-        // in this case there may be a table manifest present
-        file_put_contents(
-            $root . '/upload/file.manifest',
-            '{"primary_key": ["Id", "Name"]}'
-        );
-        $systemMetadata = [
-            'componentId' => 'testComponent',
-            'configurationId' => 'metadata-write-test',
-            'configurationRowId' => '12345',
-            'branchId' => '',
-            'runId' => '999',
-        ];
-        $tableFiles = [
-            'tags' => ['output-mapping-test', 'another-tag'],
-            'is_permanent' => true,
-        ];
-
-        $writer = new FileWriter($this->getStagingFactory());
-
-        $writer->uploadFiles(
-            '/upload',
-            [],
-            $systemMetadata,
-            StrategyFactory::LOCAL,
-            $tableFiles
-        );
-        sleep(1);
-
-        $options = new ListFilesOptions();
-        $options->setTags(["output-mapping-test"]);
-        $files = $this->clientWrapper->getBasicClient()->listFiles($options);
-        $this->assertCount(1, $files);
-
-        $expectedTags = [
-            'output-mapping-test',
-            'another-tag',
-            'componentId: testComponent',
-            'configurationId: metadata-write-test',
-            'configurationRowId: 12345',
-            'branchId: ',
-            'runId: 999',
-        ];
-        $file = $files[0];
-        $this->assertNotNull($file);
-        $this->assertEquals(4, $file['sizeBytes']);
-        $this->assertEquals(sort($expectedTags), sort($file['tags']));
-        $this->assertNull($file['maxAgeDays']);
     }
 }
